// src/types/interfaces/manageControls.interface.tsx
export interface IControlsTypes {
  control_id: number;
  control_name: string;
  description: String;
  pending_results_count: number;
  control_type: String;
  priority: number;
  datasources: String;
  control_last_run: String;
  control_last_updated: String;
  status: String;
  state: String;
  result_investigator: String;
  authorized_data: String;
  revision: number;
  revision_date: String;
  created_by: String;
  created_date: String;
}